scripts: revert change to package `git describe`
This was mistakenly not rebased out of c8994d2 ('ci: move to github
actions'). There should never be lightweight tags on the CI-bearing
remote, so --tags is either a no-op or an error in CI, and it is
always an error locally if the local user uses a lightweight tag.

--broken is a useful safeguard for weird or inconsistent repository
states, and --dirty distinguishes things which are not exactly the
commit in question.