Events endpoint should listen for and discard oracle commands
When we are running the ledger in events endpoint mode for testing
purposes, the shell is still provided with the handle to an Ethereum
oracle, though none is actually running. The events endpoint task should
listen for and discard any commands the shell sends that were intended
for the oracle.