Fix transfers to Namada e2e test
The Ethereum events queue seems to be queueing events as expected.
This test had an event being confirmed for a nonce higher than the
ledger's nonce, therefore it failed after the introduction of the
Ethereum events queue. The test now uses a nonce matching the ledger's
stored transfers to Namada nonce.