apps/config: use blank domain and organization
The application data directory standards on Windows and Mac use
organization names in the directory paths. On Windows, this is a path
like `%AppData%\OrgName\AppName`, while on Mac it is a path like
`$HOME/Library/Application Support/com.example.AppName`, using a
Java-like embedding into DNS, for Mac application bundles.

Namada does not provide a Mac application bundle, so the DNS embedding
is unnecessary, and "heliax.com" is not the correct domain name in any
case. Therefore, use an empty string for the domain qualifier.

The correct organization name for Namada might be "Namada" if one were
necessary, but it would be redundant with the application name.
Therefore, use an empty string for the organization name.

This results in paths like `%AppData\namada` for Windows and
`$HOME/Library/Application Support/namada` for Mac.

The Unix directory name, `$XDG_DATA_HOME/namada`, is unaffected.

A useful citation follows:

File System Programming Guide (for Mac):
https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/FileSystemProgrammingGuide/FileSystemOverview/FileSystemOverview.html
(see Table 1-3; note that it specifies "your app's bundle identifier or
your company" but many organizations or communities, like Namada, are
not companies; we can read this broadly to encompass "Namada", and it
is not unheard of to use a simple application name here.)