wallet: add forcible alias overwriting options
This adds WALLET_ALIAS_FORCE to the client and ALIAS_FORCE to the
wallet, forcing an overwrite of conflicting aliases without prompting
the user. They are not actually wired up to command line flags here,
but could be.

init-network and join-network, as well as genesis setup and dev build
defaults, unconditionally overwrite aliases.