scripts/build_network: remove spurious field separators
The default value of $IFS (internal field separator) is space, tab, and
newline. When de-bashifying this script, this line will change its
meaning to make the literal characters 'n' and 't' field separators,
causing heinous errors (for example, the word "network" is likely to
show up, which contains an 'n' and a 't').

But the line serves no purpose here, so remove it.