masp: implement remaining inflation logic
Inflation logic that works with 256-bit Amounts was deferred until now.
We correct the implementation to adjust rewards for words above the
lowest. (Rewards are only awarded in the lowest denomination of NAM,
which is the only one used; if a reward exceeds i64::MAX, this is
certainly an error.)

For nonstandard (non-10^6) denominations, we must also perform an
adjustment in decimal places. This will apply to ETH in reality, and
our test ETH now.