Derive Traits for address::DecodeError
These trait derivations should make it easier for code who wish
to ingest this type to themselves derive traits like Eq and Clone.

This Topic relies on #1818 as it sets the ground work for being able
to derive this information. However, this work can be done by hand
another way, and alternative strategies can be had to derive such
information