Move PinedBallancedError to error.rs and Integrate it into the type
We move PinedBalanceError into the greater error type, and properly
move over errors which may be broader into the more general error
type.

Further a creation of a helper function is made to help determine if a
function is a pinned error or not, as our greater type lacks any sort
of equality traits, and thus the old way of using != and == could not
be used.