Remove unwraps from shared::ledger::events
This commit removes all but 1 unwrap from the events file.

These unwraps were caused by potential serde errors. With the change
of the types, we are able to capture these kinds of failures, and are
able to move it towards a more accurate error type and message to
accompany it.

In this process, the encoding error type has been made. This type
should be the home of any kind of seralization and deseralization
errors we may find throughout the codebase