Make query_wasm_code_hash return result, and do not .expect results
Every call site of query_wasm_code_hash called .unwrap().await; We
instead replace this pattern with a ?; This is possible since the
error type is now general enough to encapsulate query errors.

Further there are 17 call site where unwrapping the key was had in Tx. Thus,
we've created a helper function that properly reports an error back on
the pathbuf if it can't be converted.

There are a further 13 uses in signing