Abstract out the transaction builder into a generic build function
This commit simply removes a fair amount of boilerplate around
building transaction. Before the same logic would be repeated 14
times. Which make future refactoring annoying, as the same solution
would have to be done around 14 times.

Now we've abstracted it in a generic function called build, which
prepares and builds the transaction

Test commit

Fixes the issue pointed out by yuji will be put in the proper place