Convert unwrap_client_response's into convert_response's
convert_response returns a result that can be handled with ? rather
than throwing an unrecoverbale panic.

This Commit also starts to convert many functions called to use this
function.

For functions which were discriminated on by their Optional status and
turned into an error, considerations on the error messages were made

1. If the response is not the explicit response they expected
   before (I.E. not an explicit return of None), then we make sure to
   match the specific cases the None turned into.

2. If the logic was discriminated against only in a success way, then
   the catch for general errors was added as it could cause
   unintentional bugs

This commit is mostly exhaustive, only a few invocations remain with
unwrap_client_resposne. Namely functions within a feature flag, and
one that is a bit tricky to handle