Remove unexpected events from the IBC VP
This fix is not the most robust solution, but it is the one that
involves less code changes. Ideally, we would include the wrapper tx
hash in the native VP execution context, to compute the hash of the
inner transactions that is now being included in each wasm emitted
event, but this change is too invasive.